Implement separate address translation unit
If an M-mode interrupt is taken while the SHA3 accelerator is running in
a virtual memory environment, the accelerator will fault and hang from
virtual addressing being disabled from underneath it.  Work around this
by enabling the accelerator to perform virtual address translation
independent of the current privilege level.

Since the satp CSR continues to be shared with Rocket, this approach
still relies on a fence instruction to precede each context switch.